Refactor window modules
