+++ src/builtin_plugins/amulet_team_editor/window/__init__.py
@@ -1,3 +1,3 @@
 from ._main import ButtonProxy
-from ._child import AmuletSubWindow, create_sub_window
+from ._child import AmuletChildWindow, create_sub_window
 from ._tab_engine import TabWidget
