+++ src/builtin_plugins/amulet_team_editor/window/_child/child_window_ui.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>AmuletSubWindow</class>
- <widget class="QMainWindow" name="AmuletSubWindow">
+ <class>AmuletChildWindow</class>
+ <widget class="QMainWindow" name="AmuletChildWindow">
   <property name="geometry">
    <rect>
     <x>0</x>
